Added Delete Task Feature
The user now has the option to delete a task from the task list.

First, it checks if there are any tasks to delete then displays the current tasks with numbers for reference and gets the task number to delete from the user. The deleetion happens using the slice operation